+++ tests/test_convex_adam_mind.py
@@ -2,41 +2,16 @@ from pathlib import Path
 
 import numpy as np
 import SimpleITK as sitk
+from helper_functions import (resample_img, resample_moving_to_fixed,
+                              rotate_image_around_center_affine,
+                              rotate_image_around_center_resample)
 
 from convexAdam.apply_convex import apply_convex
 from convexAdam.convex_adam_MIND import convex_adam_pt
 
 
-def resample_img(img: sitk.Image, spacing: tuple[float, float, float]) -> sitk.Image:
-    resample = sitk.ResampleImageFilter()
-    resample.SetOutputSpacing(spacing)
-    resample.SetSize([int(sz * spc / new_spc + 0.5) for sz, spc, new_spc in zip(img.GetSize(), img.GetSpacing(), spacing)])
-    resample.SetOutputDirection(img.GetDirection())
-    resample.SetOutputOrigin(img.GetOrigin())
-    resample.SetTransform(sitk.Transform())
-    resample.SetDefaultPixelValue(0)  # value for regions without source (zero-padding)
-    resample.SetInterpolator(sitk.sitkLinear)
-
-    return resample.Execute(img)
-
-
-def resample_moving_to_fixed(fixed: sitk.Image, moving: sitk.Image) -> sitk.Image:
-    """Resample moving image to the same grid as the fixed image"""
-    resample = sitk.ResampleImageFilter()
-    resample.SetOutputSpacing(fixed.GetSpacing())
-    resample.SetSize(fixed.GetSize())
-    resample.SetOutputDirection(fixed.GetDirection())
-    resample.SetOutputOrigin(fixed.GetOrigin())
-    resample.SetTransform(sitk.Transform())
-    resample.SetDefaultPixelValue(0)  # value for regions without source (zero-padding)
-    resample.SetInterpolator(sitk.sitkLinear)
-
-    return resample.Execute(moving)
-
-
 def test_convex_adam_identity(
     input_dir = Path("tests/input"),
-    output_dir = Path("tests/output"),
     subject_id = "10000_1000000",
 ):
     # paths
@@ -142,61 +117,6 @@ def test_convex_adam_translation(
 
     # compare with reference (displacement field should be within 1 mm of the translation for at least 90% of the voxels)
     assert (np.abs(displacementfield + 10) < 1).mean() > 0.9
-
-
-def rotate_image_around_center_affine(moving_image: sitk.Image, angle: float) -> None:
-    """
-    Rotate the given image around its center by the specified angle.
-
-    Parameters:
-        moving_image (sitk.image): The image to be rotated.
-        angle (float): The angle of rotation in radians.
-    """
-    # Calculate the physical center of the image
-    size = moving_image.GetSize()
-    spacing = moving_image.GetSpacing()
-    physical_center = [(sz-1)*spc/2.0 for sz, spc in zip(size, spacing)]
-
-    # For a 3D image rotation around the z-axis, the rotation matrix is:
-    rotation_matrix = np.array([[np.cos(angle), -np.sin(angle), 0],
-                                [np.sin(angle),  np.cos(angle), 0],
-                                [0,              0,             1]])
-
-    # Compute the new origin after rotation
-    new_origin = np.dot(rotation_matrix, -np.array(physical_center)) + np.array(physical_center)
-
-    # Get the current direction of the image
-    direction = np.array(moving_image.GetDirection()).reshape((3, 3))
-
-    # Compute the new direction cosines by multiplying the current direction by the rotation matrix
-    new_direction_cosines = np.dot(rotation_matrix, direction)
-
-    # Update the image with the new direction and origin
-    moving_image.SetDirection(new_direction_cosines.flatten())
-    moving_image.SetOrigin(new_origin)
-
-
-def rotate_image_around_center_resample(moving_image: sitk.Image, angle: float) -> sitk.Image:
-    """
-    Rotate the given image around its center by the specified angle.
-
-    Parameters:
-        moving_image (sitk.image): The image to be rotated.
-        angle (float): The angle of rotation in radians.
-    """
-    scale_factor = 1.0
-    translation = (0, 0, 0)
-    rotation_center = moving_image.TransformContinuousIndexToPhysicalPoint(np.array(moving_image.GetSize())/2.0)
-    axis = (0, 0, 1)
-
-    # rotate moving image
-    similarity_transform = sitk.Similarity3DTransform(
-        scale_factor, axis, angle, translation, rotation_center
-    )
-
-    moving_image = sitk.Resample(moving_image, similarity_transform)
-
-    return moving_image
 
 
 def test_convex_adam_identity_rotated_direction(
